Module20/CapstoneEvaluation/requirements.txt
python-decouple==3.8
matplotlib==3.7.3
missingno==0.5.2
numpy==1.23.4
pandas==1.5.3
plotly==5.18.0
scipy==1.10.1
seaborn==0.11.2
sklearn==1.3.2
alpaca-py==1.6.0
alpaca-traiding-api==3.2.0
finta==1.3
transformers==4.37.2